Seperated the function find_freq.c into it's own file *Modified all files so that they would reference find_freq.h *clean up
