+++ bbb/meta-bbb/recipes-core/trebletrouble/files/src/find_freq.h
@@ -1,0 +1,1 @@
+int find_freq(double freq);
